Warning fix: reinterpret_cast that changes size.
In wx.cpp I needed to store an 'unsigned' in a private-data field of
type 'void *', and did it via a reinterpret_cast. That provokes a
warning from clang, on the basis that it's confusing for
reinterpret_cast to change the size of a thing.

Easily enough fixed by first static_casting the input integer to one
the same size as the pointer, and then reinterpret_casting *that* to
an actual pointer. That explicitly expresses the semantics I wanted,
and doesn't give a warning.

+++ browser/wx.cpp
@@ -1622,8 +1622,8 @@ void SubsidiaryView::update_trace_window_list()
         if (tw) {
             ostringstream oss;
             oss << "#" << tw->window_index;
-            linkcombo->Append(oss.str(),
-                              reinterpret_cast<void *>(tw->window_index));
+            linkcombo->Append(oss.str(), reinterpret_cast<void *>(
+                                  static_cast<intptr_t>(tw->window_index)));
             if (tw == this->tw)
                 id = curr_id;
             curr_id++;
